third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/focus-shadow-crash-expected.txt

CONSOLE ERROR: line 12: Uncaught TypeError: host.createShadowRoot is not a function
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE


Pressing TAB key to navigate a shadow insertion point in the oldest shadow tree should not make the browser crash.
